requirements.txt
streamlit
fitz
json
spacy
PyMuPDF
python-docx
networkx
pygraphviz